feat(brand): the launcher is the same mark, still until you hover it
The app-launcher logo was eight vertices on eight separate CSS keyframe
tracks, each tweening its own cx/cy, plus dashed "dynamic" edges on a
ninth. It could not produce a coherent projection because nothing tied the
vertices to a shape — they were eight independent animations that happened
to start on a cube, which is why the edges drifted out of square as soon as
it ran.

It is now the real 4D projection, from the same module the loader draws
from, so the launcher and the loading state are recognisably one mark
rather than two lookalikes that will drift.

Still by default. At rest it is a logo and should behave like one: a
precise isometric hexagon that never competes with the content beside it.
Chrome that moves on its own is chrome people learn to tune out.

Under a pointer it runs the same invert-turn-invert loop over 15 seconds
against the loader's 4.5. Three times slower is the difference between
"this is alive" and "this is working" — a launcher inverting at loading
speed would read as a progress indicator and imply something was happening
when nothing was. Touch has no hover, so on a phone it is simply the static
mark, which is the right outcome rather than a gap.

Extracted `lib/brand/tesseract-geometry` (vertices, edges, isometric camera,
ZW rotation, the two schedules) and `TesseractMark` (the drawing, driven by
a period and a boolean). `TesseractLoader` is now only the loading
composition. The geometry was going to be needed twice; a second copy is
how two logos become two different logos.

No vertex nodes at launcher size — sixteen dots on a 24px mark close the
gaps between the lines and it reads as a blob — and a heavier stroke,
because a hairline at 24px disappears into the header.

The projection tests move to `lib/brand` with the geometry, and
`guard:unit` gains `src/lib/brand` and `src/components/ui`, neither of
which it covered.

unit 427 · layout 193 · card-surface 0 · ci · holdings · tdz.
